[BEAM-14157] GrpcWindmillServer: Use stream specific boolean to do client closed check (#17191)
* [BEAM-14157] GrpcWindmillServer: Use stream specific boolean to do client closed check

This is a follow up to #17162. An AbstractWindmillStream can have more than one grpc stream during its lifetime, new streams can be created after client closed for sending pending requests. So it is not correct to check `if(clientClosed)` in `send()`, this PR adds a new grpc stream level boolean to do the closed check in `send()`.

* [BEAM-14157] Add unit test testing CommitWorkStream retries around stream closing

* [BEAM-14157] review comments

* [BEAM-14157] review comments

* [BEAM-14157] review comments

* [BEAM-14157] fix test

* [BEAM-14157] fix test

Co-authored-by: Arun Pandian <pandiana@google.com>