BEAM-14277: Disables Spanner change streams tests
Disables the Spanner change streams tests for now. They are currently
failing because they are using the default auto-scaling algorithm
(throughput-based), which is not supported for change streams.

We decided to disable the tests here, because we can not change the
autoscaling algorithm without adding the dependency of dataflow into the
tests module. Since we are soon re-enabling auto-scaling, we will also
re-enable the tests when we submit the PR then.