[BEAM-14464] More efficient grouping keys in precombiner table.
The current code constructs, hashes, and compares full WindowedValues
for the grouping key, which ends up dominating the time spent in the
combining table when using trivial combiners (like Sum.integers()).
We only need compare the (structural value of the) key and windows,
and can emit the windows in the global case.

Per added microbenchmarks, this is roughly a 50% improvement for
singly-windowed values, and roughly 2.5x improvement for the common
GlobalWindows case.