Go SDK: Allow overriding the default ReStream implementation for GBK and CoGBK.
This change defines a new exec.ReStreamFactory type for constructing a
re-iterable stream from a byte stream and coder.Coder. The user may override the
default behavior by calling exec.SetReStreamFactory.

This customizable factory function allows overriding the default behavior of
exec.ReStream construction. The default implementation buffers all elements
being iterated over by a GBK or CoGBK function into an in-memory buffer, which
may cause OOM errors when a single iterable is large and the runner does not
supply a state-backed iterable. See
https://github.com/apache/beam/issues/21817. The user may specify an alternative
function that spills elements to disk if a memory threshold is exceeded, for
instance.