Go SDK: Update memfs to parse the List() pattern as a glob, not a regexp (#21943)
* Go SDK: Update memfs to parse the List() pattern as a glob, not a regexp.

This makes the memfs implementation consistent with the filesystem
implementation of List().

The docstring for filesystem.Interface.List does not specify how the pattern
should be interpretted. It says: "List expands a pattern to a list of
filenames." Perhaps that docstring should be updated to be more specific.

* Use "filepath" instead of "path" in memfs; add test cases.

Also, accept input without a "memfs://" prefix.