[#19857] Migrate to using a memory aware cache within the Python SDK harness
This relies on getting the deep object size. objsize seemed like an appropriate library to consume over the larger and more complex pympler library.

We get rid of the state cache metrics in favor of plugging in with the status client output since the state cache metrics are not defined within the monitoring.proto as a well defined type. In the future we may re-introduce a similar set of metrics.