Allow expansion service to choose pickler.
This needs to be added here as it is a global option, not a pipeline option.
This is useful as dill has some issues with objects created from exec/eval
which are used for remotely-defined PythonCallable objects.

Also refactored the pickler choice logic a bit for clarity and better
error checking.