Update Python katas to latest version of EduTools
With the latest versions of the EduTools plugin, the previous katas stopped
working, and this made it impossible to release new versions of the
katas (e.g. after a new Beam release).

The main change is in the testing method. EduTools now uses Python's unittest
for checking the exercises.

I have also re-organized some of the katas (e.g. windowing and triggers are now
under the Streaming category), and updated the katas to Beam 2.41.