fix: only report backlog bytes on data records
We were estimating backlog bytes on all types of change stream records,
but the only ones that contributed to the pipeline throughput were the
data change records. This caused an issue when throughput was 0, but the
backlog bytes were > 0 due to heartbeat records, preventing a downscale
event in autoscaling.

In this change we only report the data change records bytes in the get
size function, such that when throughput is 0, even if we were receiving
heartbeat or child partition records the backlog bytes would remain at
0.