tests: fixes SpannerIO unavailable retry test
The failing test exercises the retry mechanism of the Cloud Spanner
client when receiving UNAVAILABLE exceptions. Since this error is
retried forever, the test is configured to retry for a while and cancel
the pipeline. The wait time was too low and no requests were retried.

Here we update the wait time from 5ms to 500ms, which gives the pipeline
more time to retry requests.

Fixes #21533