Implement sibling protocol for Typescript.
Though there are various worker thread libraries in
Node.js, the lack of a shared memory model makes it
difficult to share caches, negating most of the benefit
of trying to run multiple threads in the same process.
In addition, having multiple independent workers connect
to the control service is simpler (and possibly more
efficient) than adding a proxying service within the
javascript process.