Add map_windows support to Go SDK
Adds support to the Go SDK for the [map_windows][1] urn.

The existing type model does not allow `IntervalWindow` as
a FullValue and the existing coders will throw away the window
values when decoding the KV type [here][2] in `elideSingleElmFV`.
This change updates coder construction to allow an IntervalWindow
as a value coder.

[1]: https://github.com/apache/beam/blob/master/model/pipeline/src/main/proto/org/apache/beam/model/pipeline/v1/beam_runner_api.proto#L296
[2]: https://github.com/camphillips22/beam/blob/8be4cdcaf65a1e53e3041ac3354e2e99c845e915/sdks/go/pkg/beam/core/runtime/exec/coder.go#L580-L580