Suppress keyfor warnings
Starting with checker 3.16.0 keyfor defaults for the java.lang.* package
changed in a way that requires a lot of esoteric annotations. The main use of
keyfor checking in our codebase is to avoid having to check for null in the
result of Map and Set access, etc. Instead of having annotations and type
errors that are exceedingly difficult, we can simply check for null in these
cases.