Change UnboundedScheduledExecutorService to avoid creating threads when (#25300)
there is contention on the SynchronousQueue.

ThreadPoolExecutor creates a new thread if offer(Runnable) returns false.
With a SynchronousQueue under contention, this may return false even if
there are already idle threads in the pool. To avoid creating a thread
unnecessarily and growing the pool we modify offer to wait with a small
timeout of 100ms. This may delay initial execution on the pool but that
should be rare as it is encountered only once the high-watermark is
exceeded.

Also add a timeout to the non-core threads so that the pool reduces size.

Both of these changes are motivated to reduce memory usage due to per-thread
buffers in Beam and AvroCoder.

Add a test contending adds on the pool which previously showed the pool growing
beyond the necessary amount (ie created 104 threads when 100 is sufficient).