Create script_sklearn_model
I am attaching a file which contains the script that train the housing data and saves the trained models to GCS ,I dont have an GCS account so i hve just put in the code that will save the model to Gcs location.This file just contains the code and reflect the things that it will do. I have not yet run this in pipeline .In the moment just want feedback on the code which will helpme to fix the issue.