Revise error message to state the actual size limit
When a row is between 9 to 10 MB in size, the following error  occurs and confuses users with inconsistent size limit information.

"Error message from worker: java.lang.RuntimeException: We have observed a row that is 9641699 bytes in size. BigQuery supports request sizes up to 10MB, and this row is too large.  You may change your retry strategy to unblock this pipeline, and  the row will be output as a failed insert."

Updating the error message to display the actual size limit.