Speed up `IntervalWindow` argument construction
In the original snippet, we create two instances of
`Timestamp(micros=s)`, the later is rather short
lived as we immediately use the `__add__` dunder
method to create a brand new instance of `Timestamp`.

The change in this commit is to remove the construction
of this short lived `Timestamp` and reuse the one already
constructed for the `start` argument of `IntervalWindow`.

The speed gain of this change depends on the specification
of the `SlidingWindows`. I have a pipeline that is relatively
simple and with GCP's profiler in Dataflow, it revealed about ~2.5% of
wall time is wasted in constructing this short lived `Timestamp`.

Given that `__add__` doesn't alter the `Timestamp` object inplace
(and unlikely to be changed to that behaviour), I think this is
a safe change.