Better errors when inputs are omitted.
It's not always possible to know if a transform consumes inputs,
or can act as a root transform (and in fact some may be able to do
both depending on their configuration), but when a transform
expecting inputs doesn't get them the error can be quite obscure.
This adds best-effort checking and a better error in that case.

We also allow explicitly setting empty imputs to work around this
error (which is where most of the complexity of this change lies).
Importantly, sources (no matter their name) are not required to have
inputs.