Ensure configuration schema used to decode configuration.
Just because schemas are compatible doesn't mean that they're identical (e.g. up to field ordering).

This is important as schema generation in Java is not always determanistic, so even if the payload is encoded with the exact schema that was previously provided it might not agree when the request comes in.