More efficient map side inputs for small maps.
The ability to perform point lookups for multi-map side inputs is great
for maps that are too large to fit into memory, but can be very inefficient
in requiring an entire state request per key for small maps.

This change adds an optional protocol to request an entire map as a stream
of key-values in one (possibly paginated) API call, and uses this to bulk
pre-fetch an initial set of values from the map before falling back to
point lookups.