Allow users to configure wait options for new Neo4j databases
Since v5 of Neo4j, databases are created in an asynchronous manner.
Before this commit, `Neo4jResourceManager` did not allow users
to specify whether to wait or not for the database creation (via
the corresponding Cypher syntax element).

This commit fixes that, while keeping the behavior the same as
before, by default (which is to not wait).