[YAML] Increase re-use of providers with implicitly overlapping transforms. (#30793)
We use the schema transform discovery service to determine all the
transforms that a given provider vends which may be larger than those
that were explicitly declared (e.g. the basic mapping transforms
are part of java core) and use this to expand the possible set of
transforms this provider can be used to instanciate (attaching the
appropreate renaming, etc. as needed). This can greatly increase the
chances that we can use the same provider, and hence same SDK
and environment, for adjacent steps.

This is done lazily as transforms are used to avoid instanciating
all possible providers for all pipelines.