Add in-memory variants of side inputs.
This is useful for cases where the side input may be too large for our
default caching infrastructure but the user would nonetheless prefer to
spend the memory on keeping the entire object live.

This can be especially useful for Maps, where it may be much more efficient
that doing point lookups.