Fix issue where manual subsequent label designations were ignored.
The superclass implementation of __rrshift__ was getting overridden
by the __getattr__ code that was writen with the express intent of
preserving the given label (for other operations).

In particular this is needed for transforms like ReadFromCsv where
a default label is provided but users should be able to override it.