Avoid length-prefix-bytes substitutions for Flink boundaries.
The runner infrastructure typically substitutes "unknown" coders for
length-prefixed bytes, but for elements going into or out of the
Beam pipeline fragment we want the actual objects.  Fortunately, we
know these coders are known on the Runner side as well as the SDK
side as we have them in hand there.