JCSMP properties providers for new SolaceIO write connector (#31906)
* JCSMP properties providers for new SolaceIO write connector

This adds a base class and two providers for the new SolaceIO write
connector. The basic authentication provider just uses a username and
password.

Users can write their own providers to set authentication mechanisms
and other session properties, by extending from the base class. As an
example, we provide a GoogleCloudSecretProvider, to show how to create
your own. This provider can also be used "as is", but it is mainly
meant as an example to implement your own.

In upcoming PRs, I will be submitting the rest of the write
connector. It is thousands of lines of code, so I am splitting in
smaller PRs, to facilitate code reviews.

This PR contributes to #31905

* Fix CheckStyle and Spotbugs errors

* Fix class dependencies declaration

* Unify session and auth providers/factories for the read and write connectors

* Fix format violations in comment