Add WindowedvalueParam option to DoFns.
This is useful when one wants to capture and preserve all the
windowing information for later use (and is both more convenient
and less error-prone than gathering the timestamp, windows, and
pane info separately).

New values with the same windowing value can be constructed via
WindowedValue.with_value(...) and emitted asynchronously (e.g.
after batching or in finish_bundle).