Use proper coders in interactive cache.
Formerly the coder used was always a url-excaped pickling[1]
of windowed values. This is quite inefficient in time and space.

The default (text) sink is modified to use base64 encoding to avoid
embedded newlines, and also has compression by defuault (which
helps enormously in the case of common windowing and timestamp
metadata).

[1] The FastPrimitivesCoder is targeted at efficiently coding elements,
not windows or windowedvalues.

Also add compression to account for base64 expansion and (often)
highly repetative windowing metadata.