Avoid holding data elements alive via stack frame gc roots.
This is accomplished by changing to iterators which throw away
elements that have been passed as well as just referring to
the inputstream of the element instead of the entire stream.
If the ByteString is based upon a ByteBuffer this will allow
blocks that have been advanced past to be gc'd.