[Dataflow Streaming] Support for using virtual harness threads
This is an experimental feature that can be on Java 21+ by
passing pipeline option `--virtualHarnessThreads=true`.

In Java 21, virtual threads pin the underlying platform thread
when inside synchronized blocks or in native code. When enabling this,
care should be taken by user code to not block and pin the harness threads for long.

Tested the feature on sample jobs. Will send separate changes to
optimize blocking synchronize blocks, if any, in runner code.