Run a significantly smaller Dask Graph. (#33806)
* Run a significantly smaller Dask Graph.

`dask.optimize()` isn't a cure-all for everything. It won't remove tasks in a highly-duplicated graph, like the one we collect in the DaskRunner's visitor. The fix, however, is quite simple: We can get away with the dask distributed client only running the last value in the translated operation graph.

Even with this small change, all tests pass. Further, it should lead to a significant performance improvement.

* Fixing all run_pylint.sh issues.

* *shakes fist at `isort`*.