Use KafkaSource instead of FlinkKafkaConsumer (#77)
This pull request includes significant updates to the `runners/flink`
module, primarily focusing on upgrading dependencies and refactoring the
Kafka-related code to use updated classes and methods. The most
important changes include updating the dependencies in
`flink_runner.gradle`, refactoring the
`LyftFlinkStreamingPortableTranslations` class to use the
`LyftKafkaSourceBuilder` and `KafkaSource`, and modifying the
`ByteArrayWindowedValueSchema` class to implement
`KafkaRecordDeserializationSchema`.

---------

Co-authored-by: Kyle Bilton <kbilton@lyft.com>
Co-authored-by: Kyle Bilton <21045365+kjbilton@users.noreply.github.com>