Use individual windows rather than window sets in the combining table. (#34193)
Using windows sets does not allow us to correctly compute timestamps
for the resulting element when there is more than one window.
This is particularly problematic for runners (such as Dataflow)
that base watermarks (and window firings) on these timestamps
by arranging that elements are seen in timstamp sorted order (e.g.
by using the timestamp as a secondary key).