[Go] Fix reshuffle implementation to use the global window for gbk output.
Otherwise the original non-reified windowing is used and can confuse
the runner, which encodes the gbk input using global windows but was
attempting to decode it as original windowing.

This fixes Dataflow runner harness failures when a
non-global windowed pcollection is reshuffled.  A work around before
this fix would be to explicitly rewindow to the global window before
reshuffling and reapply the desired windowing after reshuffling.

This isn't caught by prism because the entire reshuffle is replaced.