More eagerly merge compatible environments.
Specifically, merge environments that are identical except for their
dependencies, but where one environment's dependencies are a superset
of the others.

This will reduce the number of distinct environments of the same
type, and the associated materialization costs of moving between
them.

Ideally this would be done in a runner for environments that are
already co-located on the same worker, but that is a further
optimization.