Remove caching for remainders of state backed iterables.
State backed iterables are used specifically for the case where the full
iterable in question has already been deemed onerously large to fit into
memory, which is the antithesis of the kind of thing we want to cache.
In particular, for iterables too large to entirely fit into the cache,
this will evict everything else in the cache as a side effect of iterating
over them.
Also, unlike standard state (including side input) values, these iterables
are much less likely to be iterated over more than once, reducing the
value of caching them.

Using a cache here is particularly problematic when writing avro (or parquet)
files with fixed sharding as these sinks already do significant buffering of
their own, and some of their common inputs (like org.apache.avro.util.Utf8
objects) can double or triple in heald memory when certain methods are accessed
(which is not reflected in the cache sizing algorithsm that assume objects
cannot change notably in retained memory once placed in the cache).