Add option to pickler dumps() for best-effort determinism (#34698)
* Add option to pickler dumps() for best-effort determinism

The motivation for this change is Flume caches pickled code and
non-determinism breaks the caching. While making pickling fully-determinism
is infeasible, increasing the determinism is still useful due to the
increase in cache hits. Sets are a common source of non-determinism.
This change sorts set elements to provide deterministic serialization.
Because not all types provide a comparison operator, the sorting routine
implements generic element comparison logic.

See: https://github.com/apache/beam/issues/34410

* Fix linter errors

* Restore previous pickler settings and improve determinism test cases

* Restore pickler settings in finally block

* Plumb feature flag from runner to to_runner_api_pickled

The FlumeRunner will enable this feature flag via its canary mechansim.
The option is localized to only serialization of transforms because that
is what the FlumeRunner wants to cache. It will not enable best-effort
deterministic serialization for other uses of pickling.