Constrain DequeCoder type correctly, as it does not support nulls
The DequeCoder uses ArrayDeque internally, which disallows null elements.
We could switch Deque implementations, but this change is better. To quote the
JDK docs: "While Deque implementations are not strictly required to prohibit
the insertion of null elements, they are strongly encouraged to do so. Users of
any Deque implementations that do allow null elements are strongly encouraged
not to take advantage of the ability to insert nulls. This is so because null
is used as a special return value by various methods to indicated that the
deque is empty."