[Java Harness] Improve caching performance and measure overhead for cache internals. (#35632)
 Improvements to state caching to reduce weighing overhead and track cache internal memory usage.

Prior to this change, in a benchmark up to 10% of CPU was used in MemoryMeter.weigh. This was reduced by using the Weighted interface for common keys for the cache, reserving weighing for actual user values.

Additionally the cache internal overhead is measured by improving the Weigher
used to account for additional bytes for each entry. This better respects the
configured cache size.