Remove `gradle-wrapper.jar` from source tree
The ASF Release Policy mentions that a source release should not contain compiled code. Jar files, like the `gradle/wrapper/gradle-wrapper.jar` file, are compiled code and should not be included in source releases.
The Gradle build tool currently offers no way out-of-the box to run Gradle without having the wrapper jar in the source tree.

[The repository](https://github.com/snazy/gradle-wrapper-no-jar) demonstrates how to use Gradle without having the wrapper jar in the source tree.

This change adopts these changes for Apache Beam, and also
* Replace the deprecated [`gradle/gradle-build-action`](https://github.com/gradle/gradle-build-action) (repo is archived/read-only) with [`gradle/actions/setup-gradle`](https://github.com/gradle/actions/tree/main/setup-gradle)
* Remove usage of the no longer necessary [`gradle/wrapper-validation-action`](https://github.com/gradle/wrapper-validation-action) (repo is archived/read-only)